# Performance: Xim and Hnd readers