Until now fread was called for every compressed pixel (or pixel differences to be exact) in the file.
fread calls was the bottleneck of these readers (by intuition and according to callgrind).
Therefore reducing all these separate fread calls to one call reading the entire compressed buffer, should be more efficient, if the post-processing (decompression) of this buffer into separate pixels can be computed efficiently.

It took me a while to wrap my head around the binary operations, but I think this works.
The solution ended up being more simple than first expected, which is good because this may mean it will work on both big and little endian systems.

The reading time in "rtkvariantest" (of the respective reader->Update() calls) went from around **35-45ms|60 ms** to about **15ms|38ms (Hnd)** and **10ms|31ms (Xim)** on Linux-5.0.1 with gcc-8 in **Release|Debug** mode.

I have not tried it on windows yet, but I expect the gain may be even larger for NTFS.