@SimonRit 2019-03-25
Hi,
I didn't notice the 
```
#pragma omp parallel
```
but they give warnings, see the [dashboard](https://my.cdash.org/viewBuildError.php?type=1&buildid=1626021). Are they required for performances? I would trust the standard library for a reduction, e.g., `std::accumulate`. What do you think?
Thanks,
Simon