@agravgaard 2019-03-25
You're right, std::accumulate should be preferred. I have created another pull request: https://github.com/SimonRit/RTK/pull/241

I was confused about the return type of `std::valarray<>::apply()`, which wasn't accepted by `std::begin()` and `std::end()` and gave up on making it work a little to early. 
And `std::valarray<>::sum()` would overflow for `unsigned char`